contact us front + some modifies

+++ delete.php
@@ -11,6 +11,10 @@
 			if(isset($_POST['Yes'])) {
 
 				$conn = new mysqli("localhost","root","", "project");
+				if(!$con){ //maybe here we can throw an exception? instead of using die()
+					echo "connection error<br>";
+					die();
+				}
 				//echo $_SESSION['email'];
 				$sql="DELETE FROM users WHERE email ='". $_SESSION['email'] . "'";
 				$sql2="DELETE FROM orders WHERE customerID ='". $_SESSION['id'] . "'";
@@ -24,8 +28,9 @@
 					session_destroy();
 					echo "<script>window.location.href='Home.php'</script>";
 				}
+				$con->close();
     	    }
-        	if(isset($_POST['No'])) {
+        	else if(isset($_POST['No'])) {
 				echo "<script> 
 						$('#signOutModal .close').click(); 
 						window.history.go(-1);
